结构的小调整
