+++ group13/453070251/lessones02/ArrayUtil.java
@@ -1,5 +1,5 @@
 package lessones02;
-import others.Others;
+import others.PrimeGetter;
 import lessones01.LinkedListWithFloat;
 public class ArrayUtil{
 	public static void reverseArray(int[] arg_arr){
@@ -72,9 +72,9 @@ public class ArrayUtil{
 	}
 	public static int[] getPrimes(int max){
 		int i=0;
-		for(;Others.PrimeGetter.get(i)<max;i++){}
+		for(;PrimeGetter.get(i)<max;i++){}
 		int[] arr = new int[i];
-		for(i--;i>=0;i--){arr[i] = Others.PrimeGetter.get(i);}
+		for(i--;i>=0;i--){arr[i] = PrimeGetter.get(i);}
 		return arr;
 	}
 	private static int[] two_s_TimesEnd = {8,6,2,4};
@@ -84,7 +84,7 @@ public class ArrayUtil{
 			int z = (int)Math.pow(2,i) - 1;
 			int zz = z*two_s_TimesEnd[i&3];
 			if(zz%10 == 8||zz%10 == 6){
-				if(Others.PrimeGetter.isPrime(z)){
+				if(PrimeGetter.isPrime(z)){
 					int zzz = (int)Math.pow(2,i-1)*z;
 					if(zzz<arg_num&&zzz>0){
 						arr.add(zzz);
